trunk   gameserver     oops wonder how i forgot this
git-svn-id: http://svn.l2jfree.com/l2jfree@361 6ff00ba9-25cd-4355-987e-099c62bc2628

+++ trunk/L2_GameServer_c5/src/main/java/net/sf/l2j/gameserver/serverpackets/CharSelectInfo.java
@@ -248,7 +248,7 @@ public class CharSelectInfo extends ServerBasePacket
         charInfopackage.setHairColor(chardata.getInt("haircolor"));
         charInfopackage.setSex(chardata.getInt("sex"));
         
-        charInfopackage.setExp(chardata.getInt("exp"));
+        charInfopackage.setExp(chardata.getLong("exp"));
         charInfopackage.setSp(chardata.getInt("sp"));
         charInfopackage.setKarma(chardata.getInt("karma"));
         charInfopackage.setClanId(chardata.getInt("clanid"));
